+++ game/missiongenerator/aircraft/waypoints/joinpoint.py
@@ -8,7 +8,6 @@ from dcs.task import (
     OptECMUsing,
     OptFormation,
     Targets,
-    OptROE,
     SetUnlimitedFuelCommand,
 )
 
@@ -94,12 +93,6 @@ class JoinPointBuilder(PydcsWaypointBuilder):
         max_dist: float = 30.0,
         vertical_spacing: float = 2000.0,
     ) -> None:
-        if self.flight.is_helo:
-            # Make helicopters a bit more aggressive
-            waypoint.tasks.append(OptROE(value=OptROE.Values.OpenFireWeaponFree))
-        else:
-            waypoint.tasks.append(OptROE(value=OptROE.Values.OpenFire))
-
         rx = (random.random() + 0.1) * 333
         ry = feet(vertical_spacing).meters
         rz = (random.random() + 0.1) * 166 * random.choice([-1, 1])