+++ qt_ui/windows/newgame/WizardPages/QGeneratorSettings.py
@@ -98,6 +98,10 @@ class GeneratorOptions(QtWidgets.QWizardPage):
         self.registerField("ea6b_prowler", self.ea6b_prowler)
         self.hercules = QtWidgets.QCheckBox()
         self.registerField("hercules", self.hercules)
+        self.oh_6 = QtWidgets.QCheckBox()
+        self.registerField("oh_6", self.oh_6)
+        self.oh_6_vietnamassetpack = QtWidgets.QCheckBox()
+        self.registerField("oh_6_vietnamassetpack", self.oh_6_vietnamassetpack)
         self.uh_60l = QtWidgets.QCheckBox()
         self.registerField("uh_60l", self.uh_60l)
         self.f4bc_phantom = QtWidgets.QCheckBox()
@@ -196,6 +200,8 @@ class GeneratorOptions(QtWidgets.QWizardPage):
             ("Su-57 Felon (build-04)", self.su57_felon),
             ("Super Étendard (v2.5.5)", self.super_etendard),
             ("Swedish Military Assets pack (1.10)", self.swedishmilitaryassetspack),
+            ("OH-6 Cayuse (v1.2)", self.oh_6),
+            ("OH-6 Vietnam Asset Pack (v1.0)", self.oh_6_vietnamassetpack),
             ("UH-60L Black Hawk (v1.3.1)", self.uh_60l),
         ]