+++ resources/customized_payloads/OH-6A.lua
@@ -1,0 +1,153 @@
+local unitPayloads = {
+	["name"] = "OH-6A",
+	["payloads"] = {
+		[1] = {
+			["displayName"] = "Retribution BAI",
+			["name"] = "Retribution BAI",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{OH-6_M60_Door}",
+					["num"] = 11,
+				},
+				[2] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 6,
+				},
+				[3] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 7,
+				},
+				[4] = {
+					["CLSID"] = "{OH-6_M134_Minigun}",
+					["num"] = 8,
+				},
+				[5] = {
+					["CLSID"] = "{OH6_FRAG}",
+					["num"] = 5,
+				},
+				[6] = {
+					["CLSID"] = "{OH6_SMOKE_BLUE}",
+					["num"] = 4,
+				},
+				[7] = {
+					["CLSID"] = "{OH6_SMOKE_GREEN}",
+					["num"] = 3,
+				},
+				[8] = {
+					["CLSID"] = "{OH6_SMOKE_RED}",
+					["num"] = 2,
+				},
+				[9] = {
+					["CLSID"] = "{OH6_SMOKE_YELLOW}",
+					["num"] = 1,
+				},
+			},
+			["tasks"] = {
+				[1] = 17,
+				[2] = 35,
+			},
+		},
+		[2] = {
+			["displayName"] = "Retribution OCA/Aircraft",
+			["name"] = "Retribution OCA/Aircraft",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{OH6_XM158}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{OH6_XM158}",
+					["num"] = 9,
+				},
+				[3] = {
+					["CLSID"] = "{OH6_FRAG}",
+					["num"] = 5,
+				},
+				[4] = {
+					["CLSID"] = "{OH6_SMOKE_BLUE}",
+					["num"] = 4,
+				},
+				[5] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 6,
+				},
+				[6] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 7,
+				},
+				[7] = {
+					["CLSID"] = "{OH6_SMOKE_GREEN}",
+					["num"] = 3,
+				},
+				[8] = {
+					["CLSID"] = "{OH6_SMOKE_RED}",
+					["num"] = 2,
+				},
+				[9] = {
+					["CLSID"] = "{OH6_SMOKE_YELLOW}",
+					["num"] = 1,
+				},
+			},
+			["tasks"] = {
+				[1] = 35,
+				[2] = 17,
+				[3] = 31,
+				[4] = 15,
+			},
+		},
+		[3] = {
+			["displayName"] = "Retribution CAS",
+			["name"] = "Retribution CAS",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{OH-6 M60 Door}",
+					["num"] = 11,
+				},
+				[2] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 6,
+				},
+				[3] = {
+					["CLSID"] = "<CLEAN>",
+					["num"] = 7,
+				},
+				[4] = {
+					["CLSID"] = "{OH-6 M134 Minigun}",
+					["num"] = 8,
+				},
+				[5] = {
+					["CLSID"] = "{OH6_FRAG}",
+					["num"] = 5,
+				},
+				[6] = {
+					["CLSID"] = "{OH6_SMOKE_BLUE}",
+					["num"] = 4,
+				},
+				[7] = {
+					["CLSID"] = "{OH6_SMOKE_GREEN}",
+					["num"] = 3,
+				},
+				[8] = {
+					["CLSID"] = "{OH6_SMOKE_RED}",
+					["num"] = 2,
+				},
+				[9] = {
+					["CLSID"] = "{OH6_SMOKE_YELLOW}",
+					["num"] = 1,
+				},
+			},
+			["tasks"] = {
+				[1] = 17,
+				[2] = 35,
+			},
+		},
+	},
+	["tasks"] = {
+		[1] = 35,
+		[2] = 17,
+		[3] = 31,
+		[4] = 15,
+	},
+	["unitType"] = "OH-6A",
+}
+return unitPayloads